+++ packages/client/src/controllers/signIn.ts
@@ -1,24 +1,20 @@
-/* eslint-disable no-console */
 import { api } from '@/utils/api';
-import { ErrorResProps } from '@/utils/api/typings';
 
 export interface SignInProps {
   login: string;
   password: string;
 }
 
-export const signInController = async (values: SignInProps) => {
+export const signIn = async (userInfo: SignInProps) => {
   await api
     .postSignIn({
-      data: values,
+      data: userInfo,
+      // eslint-disable-next-line no-console
       onSuccess: (res) => console.log('%cResponse = ', 'color: green', res),
-      onError: (err) => {
-        const error = (err as ErrorResProps).response.data.reason;
-        console.log('%cError = ', 'color: red', error);
-      },
     })
     .then(() =>
       api.getUser({
+        // eslint-disable-next-line no-console
         onSuccess: (res) => console.log('%cResponse = ', 'color: green', res),
       })
     );
