+++ packages/client/src/controllers/signUp.ts
@@ -1,0 +1,20 @@
+import { api } from '@/utils/api';
+
+export interface SignUpProps {
+  first_name: 'string';
+  second_name: 'string';
+  login: 'string';
+  email: 'string';
+  password: 'string';
+  phone: 'string';
+}
+
+export const signUpController = async (userInfo: SignUpProps) => {
+  await api.postSignUp({
+    data: userInfo,
+    onSuccess: (res) => {
+      // eslint-disable-next-line no-console
+      console.log('User id =', res);
+    },
+  });
+};
