+++ packages/client/src/pages/registrationPage/RegistrationPage.tsx
@@ -1,15 +1,18 @@
 /* eslint-disable no-console */
-import React from 'react';
+import React, { useCallback } from 'react';
 import { Link } from 'react-router-dom';
 import { Typography, Row, Col, Form, Input, Button } from 'antd';
 import { RoutePath } from '@/utils/router/routeConfig';
 import './RegistrationPage.scss';
-import { signUpController, SignUpProps } from '@/controllers/signUpController';
+import { signUpController, SignUpProps } from '@/controllers/signUp';
 
 const { Title } = Typography;
 
 export const RegistrationPage = () => {
-  const submit = (values: SignUpProps) => signUpController(values);
+  const submit = useCallback(
+    (inputValues: SignUpProps) => signUpController(inputValues),
+    []
+  );
 
   return (
     <Row justify="center" align="middle" className="reg">
